reformat other tests, fix bug in thumb 4, thumb reformatting

+++ egg-vance/src/common/utility.h
@@ -3,11 +3,11 @@
 #include "integer.h"
 
 template<unsigned int bit>
-inline s64 twos(u64 value)
+inline s32 twos(u32 value)
 {
     if (value & 1 << bit)
     {
-        u64 mask = (1 << (bit + 1)) - 1;
+        u32 mask = (1 << (bit + 1)) - 1;
         return -1 * ((~value & mask) + 1);
     }
     return value;
